task: setup new directory
