+++ dataworkspace/dataworkspace/static/js/react/features/your-files-old/components/AddFolderPopup/index.jsx
