+++ dataworkspace/dataworkspace/static/js/react/features/your-files-old/components/DeleteObjectsPopup/index.jsx
