+++ dataworkspace/dataworkspace/static/js/react/features/your-files-old/components/PopupTypes/index.jsx
@@ -1,0 +1,7 @@
+const PopupTypes = {
+    ADD_FOLDER: 'addFolder',
+    UPLOAD_FILES: 'uploadFiles',
+    DELETE_OBJECTS: 'deleteObjects',
+};
+
+export default PopupTypes;
