+++ dataworkspace/dataworkspace/static/js/react/features/your-files-old/icons/TrashIcon.js
