+++ dataworkspace/dataworkspace/static/js/react/features/your-files-old/utils.js
